Moved Layout out
